[verified] Extract desired-state generation into deployment core
