+++ cli/internal/solo/state.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/devopsellence/cli/internal/config"
 	"github.com/devopsellence/cli/internal/state"
+	corestate "github.com/devopsellence/devopsellence/deployment-core/pkg/deploycore/desiredstate"
 )
 
 const soloStateSchemaVersion = 1
@@ -35,28 +36,9 @@ type AttachmentRecord struct {
 	NodeNames     []string `json:"node_names,omitempty"`
 }
 
-type SnapshotMetadata struct {
-	AppType    string `json:"app_type,omitempty"`
-	ConfigPath string `json:"config_path,omitempty"`
-	Project    string `json:"project,omitempty"`
-	UpdatedAt  string `json:"updated_at,omitempty"`
-}
+type SnapshotMetadata = corestate.SnapshotMetadata
 
-type DeploySnapshot struct {
-	WorkspaceRoot      string           `json:"workspace_root"`
-	WorkspaceKey       string           `json:"workspace_key"`
-	Environment        string           `json:"environment"`
-	Revision           string           `json:"revision"`
-	Image              string           `json:"image"`
-	Services           []serviceJSON    `json:"services,omitempty"`
-	ReleaseTask        *taskJSON        `json:"release_task,omitempty"`
-	ReleaseService     string           `json:"release_service,omitempty"`
-	ReleaseServiceKind string           `json:"release_service_kind,omitempty"`
-	Ingress            *ingressJSON     `json:"ingress,omitempty"`
-	IngressService     string           `json:"ingress_service,omitempty"`
-	IngressServiceKind string           `json:"ingress_service_kind,omitempty"`
-	Metadata           SnapshotMetadata `json:"metadata,omitempty"`
-}
+type DeploySnapshot = corestate.DeploySnapshot
 
 type SecretRecord struct {
 	WorkspaceRoot string `json:"workspace_root"`
